git_sort: Build history incrementally from each remote
Previously, history was built using all remotes at once. This led to some
nonintuitive churn towards the end of a sorted list that was later re-sorted
after remotes had been updated. For example, the order of recent commits
could change depending on how net-next, net and linux.git were merged
together.

This patch changes git_sort to build history incrementally from each
remote. For example, all commits in linux.git will appear before all
commits in net, which will in turn appear before all commits in net-next.
Hopefully this should be easier to understand.

The return value of git-sort is also modified to include information
showing which remote a commit was first found in.

Finally, if a repository appears to be a clone of linux.git, it must also
fetch from the other remotes listed in head_names. If a repository does not
appear to be a clone of linux.git, we simply use HEAD to sort commits,
instead of origin/master as was done previously.